requirements.txt
boto3==1.24.68
PyYAML==6.0
